AAC decoder dynamic port config
This allows the aac decoder OMX component to parse the configuration buffer
to gather the correct port settings for eaac, aac+, sample rate, and number of channels
as needed. A PortSettingsChanged event is then sent to the client to signal the need to
call OMX_GetParameters again.

Signed off by: Chris Kelly