MP3 Decoder Port Config
This allows the Mp3 decoder OMX component to read the mp3 frame header and
dynamically change the port settings as needed. If the frame header indicates
a port setting needs to be changed, the change is made and an event is sent to
the client to indicate the change.

Signed off by: Chris Kelly