AUDIO: mp3 omx decoder general improvements
This change set increases size of input and output buffers,
as well as adding a queue for input and output buffers that are received
duing a dynamic port reconfiguration. The number of input and output buffers
used is now 4 each. Also, SN flush command is bypassed if there are no buffers
received from the application. The mmu-faults that were seen were caused by
ControlCodecStop command not being completed before new buffers were queued
to the SN. This has been fixed by adding a wait condition (mutex) when stop is called.
When the stop completed ack is receieved, the wait condidion ends. This ensures the DSP
will be in a good state when normal buffer processing resumes.

Signed off by: Chris Kelly