AUDIO OMX: NB-AMR Encoder timing ehancements
Errors were caused by ControlCodecStop command not being completed before new buffers were queued
to the SN. This has been fixed by adding a wait condition (mutex) when stop is called. When the
stop is completed ack is receieved, the wait condidion ends. This ensures the DSP will be in a good
state when normal buffer processing resumes.