TI LCML:  OMX Video Dec WMV needs to map entire memory region.
OMX Video Dec WMV requires to map the entire memory region to avoid DSP MMUFAULT
messages from the DSP.  Making all the components to map the entire region will
reduce the performance originally pursued with the change to the mapping
function.

The index to queue buffers was changed to accomplish this condition.  The index
was no longer used and the name was changed just to avoid confusion with the
terminology.