OMX_AUDIO: WMA Decoder, fix for repeat same song case.
- Issue was unmasked after fix for small patterns.
- Before this case was passing because previous error forced the
component to reach deinitialization.
- Basically we added a reset for the first_buffer flag (to attach the
RCA header to the first buffer) without it, SN wasn't decoding right.
- Also, we added an extra check for EOS propagation, as sometimes there
are no pending buffers when playcompleted event is received, and since PV
needs to see an output buffer with the EOS marked (it ignores eventhandler
for EOS it seems), we attach it to the last output buffer queued if no
buffer pending is found.