OMX_AUDIO: NBAMR_Decoder, workaround fix to avoid dsp error condition on latest dsp baseline.
- Problem was caused when SN found an input buffer with (nFilledLen <
Mime frameLength) and usLastFrame flag set.
- When entering this IF at SN, in the success case dsp returns
playcompleted event, but it seems to keep the remaining output buffers,
making it impossible to stop the application (as it must receive all the
output buffers, one of the marked with EOS).
- When entering to error condition, dsp returned algo error.
- Workaround consists in setting nFilledLen to the MIME frame size on
the dummy input buffer that has a the frame marked with usLastFrame.