Moving existing memory debug functions to common location.
We have the same memory debug functions in all of our audio components
with minor differences in them.  Moving all of this to OMX_TI_Debug.h
to enforce consistency etc.  This is being done as an intermediate stage
commit for a MEMFREE and MALLOC change across the components.  These
functions were missing from some of our newer components.

Associated DR:  OMAPS00207687