TI LCML:  Add functional Debug printing.
This new printing for LCML will allow to print errors when present from the
DSP API so that the user will not need to recompile the LCML to enable debug
prints.

This will help to understand which DSP API call failed during critical
conditions.  This will help to rule a DSP related issued as ARM or Bridge side

With this change, LCML_DPRINT becomes deprecated and the remaining usages in the
G711 dec, G722 enc and G529 dec were removed as well.

New LCML debugging should be used as it is for OMX Components.