OMX Audio: WB-AMR Encoder - DTX, mem-macros and Astyle
[comments]
There is a change on the SN code to the structure used to pass the VAD Flag (DTX) sent from the app to the
algorithm
at the SN Control Function, which was causing a marshalling issue and prevented the VAD mode to work.

This also includes TI_OMX_Common.h macros for malloc/free.

And code clean-up using Astyle.

[reviewers]
- Chris Kelly
- Enrique Zertuche
- Jean Johnson

[DR]
None