Revert "Revert "Merge change I6769ac4d into donut""
The original changes in this patch were tested and the screen went black, as a result this patch was reverted.
Now the problem was solved through a kernel modification and rebuild the kernel.
Now this patch is being resubmitted again as a request from Chris Kelly.
Original patch number is 1080 and the revert patch is the 1132.

This reverts commit b1de669172b1e1437bf7e655253afa507f4ace92.