TI LCML: Remove compilation warnings.
To remove compilation warnings, the un-used variables of the DmmMap function
were removed as they will not cause any more confusion to the mapping parameters
passed to the DSP.

Other general warnings within the debug messages were removed.

The warnings related to the un-used variables in the shared API were kept so
that all the calling OMX components will not need to change at this moment.