OMX: Runtime check of QoS dependency
If the qos_dyn3430.dll64P file does not exist or the correct env variable
is not defined, then the RM will fall back to stub mode dynamically.
This will allow use of multimedia components w/o having to build separately with RM disabled.

Also, a typo is corrected which may have caused unexpected behavior
when booting in 3440 mode.