Remove Gray Frames at the Beginning of H263
Problem: When the file AV_000207_H263_CIF_30fps_2Mbps_AAC_HE_48KHz_128kbps.3gp is reproduced, some 10 gray frames appear at the beginning of the playback.

Cause: When this file is played, a port reconfiguration is done; this reconfiguration requires to concatenate the first buffer received from the app with
the secondly received (instead of sending them as they become available). This concatenation was done in an incorrect way.

Solution: Correct the offset used to jump the first frame and start writing the second frame.

Signed-off-by: Carlos Aceff <x0084298@ti.com>