This Patch Corrects the Way OMX' Parser Gets the fixed_vop_time_increment
Problem: File 1242023115622.mp4 does not play.

Cause: The MPEG4 parser is reading from the ES_Descriptor atom in the file a wrong width and height.

Solution: Correct the way the parser is reading the parameter fixed_vop_time_increment. The parser is reading in an incorrect way this parameter; this
issue is causing a misalignment between the structure in the file and the pointer that points to it (the pointer corrupts), making the parser read from
incorrect places the width and height.

Signed-off-by: Carlos Aceff <x0084298@ti.com>