OMX RM: removal of dsp load thread
The cpu load thread constantly polls the bridge QoS node to keep
current stats available as the basis of resource request desicions.
Becuase this thread is polling, the bridge handle was kept open through
the lifespan of RM process. This lead to a problem when the dsp recovery
daemon kills the RM to clean the stale bridge handle.

By removing the thread and closing the bridge handle, we are able to gracefully
recover without needing to kill and restart the RM process.

This fix should also lead to slightly improved performance since the load thread
is removed.

Signed-off-by: Chris Kelly <c-kelly@ti.com>