OMX:ER207275-128 bytes alignment common macro changes
Macros for allocation and free have been modified to take care of
the 128 byte alignment(ER). All OMX components need to ensure that
these macros are used for memory shared with lower layers (DSP bridge
and SN). There is a check in bridge driver and it expects 128 byte
memory aligned buffers/structures

Signed-off-by: Prasanna Kumar M.R.<prasanna.kumarmr@ti.com>