TI LCML:  Use 128byte alignment check on bridge
LCML has changed the logic to use a given allocation macro to have 128byte
aligned pointer.  LCML adapts to the bridge feature to check the 128 byte
alignment of all the pointers that pass through it to the SN.

OMX components need to adapt to this change by using the macros defined in
OMX_TI_Common.h header common file.

Add a check variable provided by bridge to perform the check to 128 byte
alignment when the buffer is written by the DSP only

Change-Id: Id80cd649f45c1ede3c2b362be7a1be350b1eec9e