TI LCML:  Invalidate output structs coming from DSP before read
It was found different memory corruptions when reading output structs from the
DSP.  The problem eases when Invalidating cache before actually reading the
structs.  This problem is not happening for all structures coming from the DSP
and this work around is presented to fix the problem.

A better solution should be delivered when the cause is identified.

This patch is based on 2418.

Change-Id: I348abf05bc65c8ab4600cd966cdd666bc58ae592