TI Config Parser:  Enable h264 cabac support
By removing the check for entropy on TI Config Parser the list of components at
PV level will keep consistency and the playback of cabac h264 files is possible.
The check for 720p resolution was never executed since PV was taking decision of
which OMX component to load beforehand.

D1 and lower resolutions H264 cabac files are currently not supported but if
tried TI Video Decoder will fail to play them.

Change-Id: I33bff6e15bdf309aeb3e547148bc83dcbfdc3bd0