revert peakBufferCpu to remove teh 10% margin
The 10% margin was added as a workaround to ensure we give a high
OPP for some a/v combo cases. This workaround is no longer required since
each MHz value requested from OMX components should represent a worst case estimate.

Also, the 10% margin was being added even in cases where the DSP max resources
were requested resulting in RM trying to set OPP for max+10percent, which does not exist.

Change-Id: I6d146155804acbe0a3820033335f53d0de812b1a
Signed-off-by: Chris Kelly <c-kelly@ti.com>