FIX: Play RTSP Streaming
For some contents of h.264 OpenCORE(OC) was returning an error, since
OC was not able to handle these incomplete frames. The capability flag
iOMXComponentCanHandleIncompleteFrames in the TI OMX component
was changed from false to true.

The issue was only reported for h.264 streams, after the validation of the change with
other formats that are regulary used in streaming (mpeg4 & h263), the changes was made
generic for all formats to avoid having format specific code.

The verification of the support of capabilities is done in OC at:
PVMFOMXBaseDecNode::CheckComponentCapabilities() function.

[ DR ]OMAPS00213506: H.264 contents are not normally playing in RTSP Streming(google streaming).

Change-Id: I6aefb873970ab22d6d13ad88d23fdab1790d088e
Signed-off-by: Vinay Shekarappa <vinay.s@ti.com>
Signed-off-by: Jorge E Solano <jsolano@ti.com>