TI OMX Core:  Increase the instances for 720p
The number on instances of 720p were limited to provide more robustness to the
encodeing path.  We have now a robust enough 720p components that have made more
evident side effects of restricting 720p to 1.

A better user experience is shown when 2 instances are possible as the following
scenarios are covered:

Record-playback-Record:  When transitioning to the playback screen a 720p dec
instance is launched to generate the big thumbnail.  If the play button is
tapped in the middle playback is not possible.  This is covered with 2
instances.

Multiple files in gallery:  When there are plenty 720p clips on SDCard, the
thumbnails will generate one by one and if in the middle of this a video is
going to be played there's a need to have the second instance that can play the
clip fine.

Robustness of recording should not be regressed by this change.

Change-Id: If4b5192ea93f0f6868f6b32094642f645b4156ce