OMX LCML: Call DSPManager_Close() properly so DSP recovery can occur
1. Remove call to free hInterface->dspCodec->hNode. At the time on a
MMU fault, hNode is already invalid. Also, hNode is kernel mapped memory,
so I don't think it should be freed in userspace. This free routinely
causes a SIGSEGV.
2. Remove DSPManager_Open() call to check for MMU fault. If MMU fault
occurs right after this call return true, DSPManager_Open() will be
incremented unnecessarily.
3. Always call DspManager_Close() in DeleteDspResource(). The call to
this function is essential for DSP to recovery properly. This function
will release the file descriptor to DSP/Bridge driver so that it can
recover. Keeping a counter for DSPManager_Open/Close() to ensure a 1:1
ratio.

Change-Id: Id5040875a100e7a1cb917e7829592b30128554f4