RM Proxy freeResources list terminated
When freeing the component from the list of registered
OMX components, the list is shifted up. This left a copy
of the last component at the last position in the list which was
not cleaned up. The fix is to set the new end of list to NULL to
effectively terminate the list of registered components.

Change-Id: I9a327e72588679198633b51bc2fdcded3aa9942f
Signed-off-by: Chris Kelly <c-kelly@ti.com>