TI Android Makefiles:  Remove linkage not needed
When linking shared objects we need to keep it to a minimum and just link
against those libraries that are actually required.  PV loads the shared object
for omx_core and the dependency of each component to it should be removed.  If
changes in omx_core should trigger a new build for the components, user should
take care of those.

Change-Id: I3f145a9979e71620b383d40a74dcc12159a1aa97
Signed-off-by: Barry Woodward <b-woodward@ti.com>
Signed-off-by: Daniel Charles <dcharles@ti.com>