OMX VD: Using stagefright in/out buffer size is defined by the client
Stagefright define the buffer sizes and OMX component must be configured
using with them.

This is not the standard accourding to OMX documentation but this way
of configure the buffer size is used by other OMX component vendors.

Memory failures will happen if the OMX component use a diferent value.

[DR] OMAPS000215074

Change-Id: I81a766cb5d8a415049300717ebc68ec1b5837678
Signed-off-by: dnkmohan <a0393586@ti.com>
Signed-off-by: Jorge E. Solano <jsolano@ti.com>