OMX Video decoder - Buffer map/unmap optimization
This patch has changes to avoid buffer mapping and unmapping done each
time at LCML level. Hence the same buffer address would be used which
helps to improve the performance. There is improvement in ARM load by
around 15MHz for MPEG4 WVGA. We continue to use MapBufLen for WMV and
the change applies to other formats

[DR] OMAPS00222176

Change-Id: I0fc1f27c2b6fbe9a894965a449f1b6c258b32b18
Signed-off-by: Prasanna Kumar M.R. <prasanna.kumarmr@ti.com>