Flush buffer on the input side needs to be optimized
It was found that we were flushing the entire input buffer
based on allocation lenght. This api use to take 9 to 12 msec.
Due to this reason the thread was getting blocked in EmptyThisBuffer
call.

We need to flush only valid data and not the allocated length.

This is helping to do precise posting of video playback frames.

Change-Id: Iee8913be8020d34d959a6997f68b30efdd7720eb
Signed-off-by: Dandawate Saket <dsaket@ti.com>